# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "EXPRESS AND PRISMA RESTful AIP EXAMPLE",
        "description": "This is a example of a backend restful api application using mainly express and prisma as its database, the repo was made by Caio Quintas, former backend programmer and tech analyst.\n\nSome useful links:\n- [Project Repo](https://github.com/Quinntas/ddd-store-client)\n- [Caio Quintas's github](https://github.com/Quinntas)",
        "contact": {
            "email": "caioquintassantiago@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/Quinntas/ddd-store-client/blob/main/LICENSE"
        },
        "version": "1.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:port/api/v1/"
        }
    ],
    "paths": {
...
        "/transaction/{publicId}": {
            "put": {
                "tags": [
                    "transaction"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Update an existing transaction",
                "description": "Update an existing transaction by publicId",
                "operationId": "updateTransaction",
                "requestBody": {
                    "description": "Update an existent transaction in database",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransactionPayload"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation exception"
                    },
                    "401": {
                        "description": "Not authorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "get": {
                "tags": [
                    "transaction"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Get an existing transaction",
                "description": "Get an existing transaction by publicId",
                "operationId": "getTransaction",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation exception"
                    },
                    "401": {
                        "description": "Not authorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "transaction"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Deletes a transaction",
                "description": "delete a transaction",
                "operationId": "deleteTransaction",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation exception"
                    },
                    "401": {
                        "description": "Not authorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/transaction/list": {
            "get": {
                "tags": [
                    "transaction"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "List all transactions for a shopkeeper",
                "description": "List all transactions for a shopkeeper",
                "operationId": "getTransactionList",
                "parameters": [
                    {
                        "name": "isTransactionAuthorized",
                        "in": "query",
                        "description": "Returns either authorized or non authorized transactions",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transaction"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation exception"
                    },
                    "401": {
                        "description": "Not authorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/transaction/authorize/{publicId}": {
            "patch": {
                "tags": [
                    "transaction"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Authorize a transaction",
                "description": "Authorize a transaction",
                "operationId": "authorizeTransaction",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation exception"
                    },
                    "401": {
                        "description": "Not authorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/transaction": {
            "post": {
                "tags": [
                    "transaction"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Add a new transaction to the database",
                "description": "Add a new transaction to the database",
                "operationId": "addTransaction",
                "requestBody": {
                    "description": "Create a new transaction",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransactionPayload"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation exception"
                    },
                    "401": {
                        "description": "Not authorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Transaction": {
                "required": [
                    "isAuthorized",
                    "amount",
                    "client",
                    "shopkeeper"
                ],
                "type": "object",
                "properties": {
                    "publicId": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "isAuthorized": {
                        "type": "boolean",
                        "example": true
                    },
                    "amount": {
                        "type": "integer",
                        "example": 10
                    },
                    "client": {
                        "$ref": "#/components/schemas/Client"
                    },
                    "shopkeeper": {
                        "$ref": "#/components/schemas/Shopkeeper"
                    }
                }
            },
            "TransactionPayload": {
                "required": [
                    "amount",
                    "clientPublicId",
                    "shopkeeperpublicId"
                ],
                "type": "object",
                "properties": {
                    "publicId": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "amount": {
                        "type": "integer",
                        "example": 10
                    },
                    "clientPublicId": {
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "shopkeeperPublicId": {
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    }
                }
            },
            "User": {
                "required": [
                    "email",
                    "name",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "publicId": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "name": {
                        "type": "string",
                        "example": "Caio Quintas Santiago"
                    },
                    "email": {
                        "type": "string",
                        "example": "caioquintassantiago@gmail.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "hmac$dasionda$ionsdsdijasdsacjdsndasdwcdn"
                    }
                }
            },
            "Wallet": {
                "required": [
                    "currentBalance"
                ],
                "type": "object",
                "properties": {
                    "publicId": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "currentBalance": {
                        "type": "integer",
                        "example": 10
                    }
                }
            },
...
            "Client": {
                "required": [
                    "cpf"
                ],
                "type": "object",
                "properties": {
                    "publicId": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "cpf": {
                        "type": "string",
                        "example": 23472594685
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "wallet": {
                        "$ref": "#/components/schemas/Wallet"
                    }
                }
            },
...
            "Shopkeeper": {
                "required": [
                    "cnpj"
                ],
                "type": "object",
                "properties": {
                    "publicId": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "cnpj": {
                        "type": "string",
                        "example": 23472594685
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "wallet": {
                        "$ref": "#/components/schemas/Wallet"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}